Replace `ExceptionMiddleware` to `ExceptionResponder`
