+++ src/Http/ExceptionResponderFactory.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Http;
+
+use Psr\Http\Message\ResponseFactoryInterface;
+use Psr\Http\Message\ResponseInterface;
+use Throwable;
+use Yiisoft\ErrorHandler\Exception\UserException;
+use Yiisoft\ErrorHandler\Middleware\ExceptionResponder;
+use Yiisoft\Injector\Injector;
+use Yiisoft\Input\Http\InputValidationException;
+
+final readonly class ExceptionResponderFactory
+{
+    public function __construct(
+        private ResponseFactoryInterface $responseFactory,
+        private ApiResponseFactory $apiResponseFactory,
+        private Injector $injector,
+    ) {
+    }
+
+    public function create(): ExceptionResponder
+    {
+        return new ExceptionResponder(
+            [
+                InputValidationException::class => $this->inputValidationException(...),
+                Throwable::class => $this->throwable(...),
+            ],
+            $this->responseFactory,
+            $this->injector,
+        );
+    }
+
+    private function inputValidationException(InputValidationException $exception): ResponseInterface
+    {
+        return $this->apiResponseFactory->failValidation($exception->getResult());
+    }
+
+    private function throwable(Throwable $exception): ResponseInterface
+    {
+        if (UserException::isUserException($exception)) {
+            $code = $exception->getCode();
+            return $this->apiResponseFactory->fail(
+                $exception->getMessage(),
+                code: is_int($code) ? $code : null,
+            );
+        }
+        throw $exception;
+    }
+}
